work in progress

+++ site/html/web/functions/delete.php
@@ -21,14 +21,14 @@ include_once("/usr/share/nginx/html/web/functions/database.php");
 $bdd = new database();
 
 //Condition de suppresion des données shouaitées pointant sur le formulaire admin pour le produit
-if ((isset($_POST['deleteMessageTab'])) && (!empty($_POST['deleteMessageTab']))) {
+if ((isset($_POST['deleteMessageTab'])) && (!empty($_POST['deleteMessageTab'])) && ($bdd->verifyUserMessage($_POST['deleteMessageTab']) == true)) {
     $bdd->deleteMessage($_POST['deleteMessageTab']);
     header("Location: http://localhost:8080/index.php?page=messages");
-} else if (isset($_POST['deleteUserTab'])) //Suppression d'autorisation de couleur pour le modèle d'un bracelet
+} else if (isset($_POST['deleteUserTab'])) 
 {
     $ret = $bdd->deleteUser($_POST['deleteUserTab']);
     header("Location: http://localhost:8080/index.php?page=users");
-} else {
-    header("Location: http://localhost/WatchOut/");
+} else{
+    header("Location: http://localhost:8080/index.php?page=messages");
 }
 ?>
